requirements.txt
numpy==2.0.1
pandas==2.2.2
plotly==5.24.1
pyarrow==17.0.0
streamlit==1.37.1
yfinance==0.2.43
pymongo==4.10.1
requests==2.32.3
pyyaml==6.0.2
google-generativeai==0.8.3